Tidy up export modal and update to be a detail popover
