+++ packages/builder/src/components/backend/DataTable/buttons/grid/GridExportButton.svelte
@@ -9,21 +9,13 @@
   $: selectedRowArray = Object.keys($selectedRows).map(id => ({ _id: id }))
 </script>
 
-<span data-ignore-click-outside="true">
-  <ExportButton
-    {disabled}
-    view={$datasource.tableId}
-    filters={$filter}
-    sorting={{
-      sortColumn: $sort.column,
-      sortOrder: $sort.order,
-    }}
-    selectedRows={selectedRowArray}
-  />
-</span>
-
-<style>
-  span {
-    display: contents;
-  }
-</style>
+<ExportButton
+  {disabled}
+  view={$datasource.tableId}
+  filters={$filter}
+  sorting={{
+    sortColumn: $sort.column,
+    sortOrder: $sort.order,
+  }}
+  selectedRows={selectedRowArray}
+/>
